Don't show '@: You'; show bars even if value is 0
The '@: You' wasn't worth keeping because it didn't even show the
current tile that the player was displayed with (difficult, because
whether the player could be seen by guards would have to be calculated
again in _draw_infopanel()) and anyway it was a brogue ripoff.

Not showing bars when the value is zero caused problems when a mob had
lost all its health but hadn't died yet (i.e., it's turn hadn't finished
yet so it wasn't marked as dead yet), as the health bar wouldn't show
even though the mob was around. Also, the `walking`/`running` bar
wouldn't show if the player wasn't doing anything, which is dumb.