Ensure that player cannot be stabbed.
Ensure that player is always aware of attacks. Stabs are there in the
first place to "reward" the player for catching a hostile off guard, but
allowing enemies to stab a paralyzed player is a bit draconian. It
shouldn't be too difficult to for player to recover from a few mistakes.