fix: don't add map into player's memory
Doing so would make the singing ring obsolete, and deprives player of a
means of knowing what portions of the map were already explored.

The "just walls and floors"-type memory that this particular commit
reverts is fine, but quite ugly and again, makes the singing ring
obsolete.

Wow, this commit message is particularly poorly written.