Revert "fix: delete disabled machines outright ..."
This reverts commit 90c15348821946a0967d124937f452f691a5fcb5.

Strange segmentation faults showed up when this commit was made and a
machine tried to delete itself. Apparently tickMachines() was still
trying to run its power function.

I have a hunch that List.remove() is the culprit, not completely
removing the node, but I'll have to investigate. For now reverting it
works perfectly.