feat: make experiments wander (ala executioners)
The whole point of experiments was to make LAB a bit more challenging.
That goal wasn't much served though when they just sit still in a pile
of vials, and are ridiculously easy to kill.

They now serve a similar purpose as the executioner, moving around the
room and difficult to stab. Just as the executioner carries an
incredibly heavy weapon, capable of two-shotting the player, the
experiments have certain capabilities that made them very dangerous to
mess with.