feat: remove rifling
In the past there was the understanding that corpses might have
something valuable to the player, and the player should rifle certain
corpses looking for interesting items.

However, this is tedious and unfun, and it clutters the screen too. Plus
the above is no longer true, mobs always carry heavy armor, clumsy
weapons, that player wouldn't want to use at all.

The only reason this should be reverted in the future is if mobs start
carrying evocables, runes, or rings that the player might want. But why
not just have that mob guard the room the evocable/rune/ring is in?

Even then, we could just add more items to the pickup menu when standing
over a corpse. Why need a rifle action at all?