fix: fix crash when kill()'ing mob in mapgen
When a mob is kill()'ed in mapgen, before the player has been
initialized, this crashes because kill() tries to check if the player
was the one responsible for killing the mob.

To fix this, kill() was split into kill() and deinit(), with kill()
doing the cause-of-death-analysis and deinit() doing the actual
deinitialization. Mapgen can now simply just call deinit() instead of
kill().