fix: mapgen: fix horrible, bad, no very good bug
This bug has apparently been the cause of mapgen forever marking a map
invalid and regenerating it endlessly. I took at look at the map, and
behold, the invalid maps were just the starting maps plopped in
different places -- no other rooms were being placed.

The cause: createCorridor() was preemptively marking the connectors for
prefabs as used, even if that corridor might be thrown away and not
executed.

Dumb me. What a pain this was to debug.