feat: optimize Coord.move
This does seem to make the game less laggy. It's probably just my
imagination though.

(I *did* test this with the old implementation w/ Hyperfine, and there is a
small speedup.)

Basically I removed a bunch of redundant checks, removing a bunch of
branching. That also allowed me to remove signed arithmetic, so only
unsigned math is done.