feat: auto-attack: revert move-to-stabbed-foes-tile change
Doesn't work well with groups of enemies, because one of them is bound
to see you when you replace their friend. Which means the game must warn
the player "Do you really want to step into an enemy's vision?" Which
raises the question, why not always let the player choose when to move
to the foe's location? Too messy.