feat: remove platform terrain
I think I might have gone overboard with adding platform back when
implementing terrain: platforms make the screen more noisy and add very
little to the gameplay. There's just not enough reason to use it to
justify its existance, as it's nearly useless without the player using
missiles.

This may be brought back as a special prefab-specific dungeon feature in
the future.