fix: spires no longer leave walls
Leaving walls is OK for statues (whose placement is highly controlled),
but bad for spires. Reason: I just encountered a spire in a doorway,
upon killing it to enter the room, the corpse swapped with me, meaning I
was standing in a wall before moving out.

In the future spires will leave heaps of tangled metal for cleaners to
remove (and replace? hmm).