fix: spells: make misses actually miss
Previously if the alternate angle of a missed bolt didn't end up
changing the trajectory, the mob would still be hit. Now if it's a miss,
anything along the trajectory will not be hit at all.

Unfortunately this means that other mobs cannot be hit "accidentally".
Which is okay, because this isn't cogmind