fix: more windows stuff
Problem: executable would crash immediately. There was a libPNG error
(when loading the font) that would result in ui.deinit() being called,
except the UI hadn't been deinited, so that would result in a
segmentation fault.

Solution: don't deinit the UI if it hasn't been inited, and change the
way fonts are read.

Apparently (??) (huge question mark, since I got this from chat gippity)
there can be issues when reading file across CRT boundaries. On Windows
there's one CRT used by the executable, and another that the libPNG dll
uses. So the usage of C apis to read the file is probably what results
in the Read Error.

In any case, when using Zig's std.fs to read the file and providing a
IO callback to libPNG, the issues go away.